+++ packages/astro-angular/vite.config.ts
@@ -5,7 +5,7 @@ import { defineConfig } from 'vite';
 // https://vitejs.dev/config/
 export default defineConfig(({ mode }) => {
   return {
-    root: __dirname,
+    root: import.meta.dirname,
     esbuild: false,
     resolve: {
       mainFields: ['module'],