+++ vitest.workspace.ts
@@ -1,4 +1,0 @@
-export default [
-  'packages/**/vite.config.{mjs,js,ts,mts}',
-  '!**/create-analog/**',
-];